Implement PulsarObject
