+++ pulsar-client-api/src/main/java/org/apache/pulsar/client/api/Schema.java
@@ -28,6 +28,7 @@ import java.time.LocalTime;
 import java.util.Date;
 import org.apache.pulsar.client.api.schema.GenericRecord;
 import org.apache.pulsar.client.api.schema.GenericSchema;
+import org.apache.pulsar.client.api.schema.PulsarObject;
 import org.apache.pulsar.client.api.schema.SchemaDefinition;
 import org.apache.pulsar.client.api.schema.SchemaInfoProvider;
 import org.apache.pulsar.client.internal.DefaultImplementation;
@@ -361,18 +362,13 @@ public interface Schema<T> extends Cloneable{
      * Create a schema instance that automatically deserialize messages
      * based on the current topic schema.
      *
-     * <p>The messages values are deserialized into a {@link GenericRecord} object.
-     *
-     * <p>Currently this is only supported with Avro and JSON schema types.
+     * <p>The messages values are deserialized into a {@link GenericRecord} object,
+     * that extends the {@link PulsarObject} interface.
      *
      * @return the auto schema instance
      */
     static Schema<GenericRecord> AUTO_CONSUME() {
         return DefaultImplementation.newAutoConsumeSchema();
-    }
-
-    static Schema<Object> OBJECT() {
-        return DefaultImplementation.newObjectSchema();
     }
 
     /**
