+++ pulsar-client/src/main/java/org/apache/pulsar/client/impl/schema/generic/GenericAvroRecord.java
@@ -67,7 +67,7 @@ public class GenericAvroRecord extends VersionedGenericRecord {
     }
 
     @Override
-    public Object getNativeRecord() {
+    public Object getNativeObject() {
         return record;
     }
