+++ pulsar-client/src/main/java/org/apache/pulsar/client/impl/schema/generic/GenericProtobufNativeRecord.java
@@ -46,7 +46,7 @@ public class GenericProtobufNativeRecord extends VersionedGenericRecord {
     }
 
     @Override
-    public Object getNativeRecord() {
+    public Object getNativeObject() {
         return record;
     }
