+++ pulsar-client/src/test/java/org/apache/pulsar/client/impl/schema/generic/GenericAvroRecordTest.java
@@ -40,7 +40,7 @@ public class GenericAvroRecordTest {
 
         GenericRecord record = schema.newRecordBuilder().set("test", "foo").build();
         assertEquals(GenericAvroRecord.class, record.getClass());
-        org.apache.avro.generic.GenericRecord nativeRecord = (org.apache.avro.generic.GenericRecord) record.getNativeRecord();
+        org.apache.avro.generic.GenericRecord nativeRecord = (org.apache.avro.generic.GenericRecord) record.getNativeObject();
         assertEquals("foo", nativeRecord.get("test").toString());
         assertEquals(1, nativeRecord.getSchema().getFields().size());
         assertEquals(schemaType, record.getSchemaType());
